Add test for exploitability rule with duplicate issue handles
Summary: Adds a test for exploitability where the exported method has multiple call-chain paths to the same underlying source/sink issue. In such cases, we want to emit a single issue to preserve the issue reporting before the exploitability rules was a thing. That is why we use set the `Issue::callee_` field as `ExploitabilityOrigin::issue_handle_callee()` which is the concatenation of the `exploitability_root` and the `callee` which partially fulfilled the exploitability rule. But since `Position` is also a part of the `Issue::GroupHash` and we currently set that to the position of the fulfilled exploitability-rule and not that of the exploitability-root, we generate 2 issues for this case in the json output. However, the position is ignored in the sapp issue handle, leading to the creation of duplicate issue handle.

Reviewed By: arthaud

Differential Revision: D62900483

fbshipit-source-id: 67c6503ac3689a799ea6a2398994a3d3bdc578c5