Remove instructions and port from positions when augmenting positions
Summary:
We store `IRInstruction` and `Root` port in `Position`. This is done so that we can compute the highlights for specific ports as a post-processing step and has the  advantages of:
- only process each file once
- only compute the specific highlight for models with issues.

But `IRInstruction` and `Root` are not a part of the json output which means that there can be duplicate positions emitted in the json which only differ in instruction and port. This diff updates the final model to include `Position` without the instruction and port fields so that we are consistent in the mariana-trench model and json output.

Reviewed By: arthaud

Differential Revision: D63039173

fbshipit-source-id: 69b5ac871ae24993e4b87a280c662b35f62fddb0