ClassIntervals: Use int32_t type for compatibility with mySQL int types
Summary:
As the title says. In the SAPP DB, trace_frames use "int" for class intervals which typically corresponds to `int32_t`.

This changes the type for class intervals from `uint32_t` to `int32_t` for consistency. This also means that class intervals can start from 0 instead of 1 (similar to the convention used in other tools).

In practice, there are not that many classes for the largest class interval to even exceed `INT32_MAX`. Overflow only occurs with `top()` intervals (`[0, UINT32_MAX]`). Behavior in SAPP depends on how the underlying DB type handles overflows.

Reviewed By: anwesht

Differential Revision: D63275431

fbshipit-source-id: c4ea4bd329d7754393b3e55876a977aa84c6c234