Create effect source for Classes with DFA annotations
Summary:
Sources exported through the Manifest can have every Intent extra attacker controlled. When a class is exported through DFA, example,
https://www.internalfb.com/code/fbsource/[c399f7a2de139677a650bc2fdd21e02510214002]/fbandroid/java/com/instagram/urlhandlers/emailconfirm/EmailConfirmExternalUrlHandlerActivity.java?lines=60-66
The only parts that are attacker controlled of the URL are the parameters that are in the template of the UriPattern.

Therefore we want superset of attacker controlled data which is `ExportedOrDfaComponent` while also leaving the subset `ExportedComponent` for example when we try and detect if a Serialized class is sent via an Intent (which is impossible over DFA which only allows strings).

In addition by switching from a feature (`via-public-dfa-scheme`) to effect sources meaning that a switch of accessScope would show up in diff analysis.

Reviewed By: anwesht

Differential Revision: D62640787

fbshipit-source-id: 67d37b3ba631747baf149b71df09d8776b2f48f1