Add points-to set using aliasing properties
Summary: This diff adds a `PointsToSet` domain. This domain will be used as an element of the PointsToTree in the following diffs. `PointsToSet` tracks a set of target memory locations that an alias may point-to. This is implemented as a Map from `MemoryLocation` to `AliasingProperties` where the `AliasingProperties` provides additional information that needs to be added when reading the memorylocation through this alias.

Reviewed By: arthaud

Differential Revision: D65186106

fbshipit-source-id: ab6fc067b48bb9064aa542cf54c26953eb3e625f