Typed exceptions
Summary:
Add a subclass of `RedexException` for the `INVALID_DEX` case. Add a flag to the debug header to switch throwing typed exceptions.

Typed exceptions allow better filtering of exceptions with the stack intact. At the same time, our current facilities for symbolication rely on the boost wrapper to attach stack information. Thus a flag to change the behavior when it suits the use case better.

Reviewed By: ssj933

Differential Revision: D66428537

fbshipit-source-id: 3f9ec7b4f7bbc9a76d6e913b40f061a1709cb044