marianabench: Emit isFalse[Negative|Positive] in annotation output
Summary:
As the title says.

Our annotations have supported `isFalse[Positive|Negative] = [true|false]` for a while now. This change enables the `ModelValidators` to parse, store and print it.

This is stored in the base class as `ModelValidator::is_false_classification`, where classification means either FP or FN, depending on the type of validator.

It is now emitted as part of the annotation string (`show()`).

Created `IssueProperties` which both `Expect[No]Issue` validators now contain.
  * Previously, these were stored in `ExpectIssue` and `ExpectNoIssue` simply contained `ExpectIssue`. Now, they each contain `IssueProperties` and are independent of each other.
  * Having `ExpectNoIssue` contain another object with its own is_false_classification flag didn't feel right, hence this change.
  * `ExpectNoIssue` is still a negation of `ExpectIssue`.

Reviewed By: arthaud

Differential Revision: D66720959

fbshipit-source-id: 255fd2cea56088bd8232ba0b2c4632d8c52d8c87