EventLogger: Support logging double values
Summary:
Noticed that our averages aren't being logged as double because the API is using "int".

Added event logging API for double values.

Also, removed the default values to `EventLogger::log(...)` since the compiler was struggling to resolve callsites to the right override. Example:
```
#1 EventLogger::log(string, string, int verbosity_level = 1);
  => EventLogger::log("event", "msg"); // This resolves to #1

#2 EventLogger::log(string, string, int value, int verbosity_level = 1);
 => EventLogger::log("event", "msg", intVal); // This can be #1 or #2
```

To differentiate callsites between `int` and `double`... any place that logs a `size_t` type now needs a `static_cast<int>(...)`.

This generally means updating a lot of call-sites, which is annoying, but probably for the better. Default values can be error prone anyway.

Reviewed By: arthaud

Differential Revision: D68037602

fbshipit-source-id: 05d25705b7ca47009910a4656c4b2d4100a8ebef