Log call graph shim metrics
Summary:
Log call graph shim metrics as well.

This is similar to the virtual callsite metrics, but looks at artificial_callees instead.

Basically, for both `stats_per_[virtual|artificial]_callsite`, we compute and log:
* Total num callsites.
* Average num resolved targets per callsite.
* P50, P90, P99 num resolved targets per callsite.

This should give a better sense of the callsite fanout problem. It can also be used to measure improvements to type precision. In general, the more precise the type information, the lower the numbers should be.

Reviewed By: anwesht

Differential Revision: D67986821

fbshipit-source-id: 999367822dcfaf3cc37cf8630acd4108603b99ff