CallGraphStats: Compute Percentage Exceeding Threshold
Summary:
We have a large skew in the histogram. However, any callsite with > 40 (join_override_threshold) targets is essentially meaningless with the JoinOverrideGenerator.

On top of knowing the P90/P99, it would be nice to know the percentage of callsites that exceed this threshold. This diff adds a percentage_above_threshold stat.

Note that this number is rather meaningless for artificial callees (see comments).

Also, since we changed types to `size_t` previously, I've updated the histogram to store `size_t` as well so that we are comparing values with the same sign-ness. They get casted back to int just before logging though, since that's all our logger supports.

Reviewed By: anwesht

Differential Revision: D68031444

fbshipit-source-id: efd0fab940a35b2a523ce7dbd7bf8a58b3ab8f6e