Add --dump-gta-call-graph option for debugging
Summary:
Adds an option to --dump-gta-call-graph, which is disabled by default.

When investigating anything GTA-related, the first question is, "is the method reachable?". The call graph can answer that. The call graph is generated by performing a DFS from the dummy "ENTRY" node.

The output is a single non-sharded .json. It is more convenient to navigate within a single file, but this convenience only applies to very small apps (e.g. marianabench).

Use this option at your own risk for larger apps. Or shard the file if the need arises.

Reviewed By: anwesht

Differential Revision: D68118305

fbshipit-source-id: 97f5496c4a6e9b62b80688c1498d3d8ba4ca659a