GTA: Log inferred method return types
Summary:
More logging for GTA to respect --log-method-types option.

Unlike `MethodContext` which stores a `dump_` flag that works per-method,  the analyzers in GTA use a shared state across methods (e.g. `WholeProgramState`). The analysis is replayed to obtain register types per instruction, so the state needs to be persisted till then.

Storing per-method state for GTA is tricky. This diff opts for simpler but inefficient ways to respect the --log-method-types option:

* Instead of logging within the GTA analyzers, log after replaying the analysis for each instruction. (In Types.cpp)
* The inferred return types are in `WholeProgramState`. Existing code dumps it but the full state is unreadable. Instead, dump only the types for methods that match --log-method-types. Inefficient but only runs if --log-method-types and --gta-verbosity 2 are specified.

On the side: Adds a --gta-verbosity flag to the shim to enable/disable GTA logs independent of MT logs.

Reviewed By: anwesht

Differential Revision: D68176800

fbshipit-source-id: 6b33250dff446bb2ad59b1abd143b541caf0747f