Log TimeoutError (and make obscure) if intraproc analysis takes too long
Summary:
Adds a timeout checks within each analyzer. The existing timer is moved into `*Fixpoint` and default constructed.

The frequency of checks must not hurt performance, but it also cannot be checked where an inner-loop/fixpoint non-termination could occur. I ended up choosing `analyze_node()`. Would be nice to check less frequently, but that would be in `MonotonicFixpointIterator`...

If a timeout occurs, the method is subsequently treated as if it were obscure. This approach does not differentiate between fixpoint non-termination vs. method being expensive to analyze.

Events are logged if timeout occurs and we should monitor them for anything unreasonable.

Reviewed By: anwesht

Differential Revision: D69096436

fbshipit-source-id: 5f7d1e0daf794e391239f2121046bf12b62ea4ca