GTA: Mark uninitialized fields' type as nullable instead of top
Summary:
Existing GTA behavior: If a static/instance field is unintialized in the constructor (both init and clinit), its type is set to top.

This is too conservative for Mariana Trench. We are reverting to an older behavior in which uninitialized fields are set to null. Its types are `NONE` at this point. Any assignment to the field results in a join with the type of the target value (as opposed to joining with top() which still gives us top()).

Reviewed By: anwesht

Differential Revision: D69691555

fbshipit-source-id: 25e80c4b7bd3752416ec216c6ba90b96836a5464