Add post order visitor for abstract tree domain
Summary: From previous diffs, we can see that the `CollapseDepth` in `AliasingProperties` is only added to a top level node. Also note that, unlike the taint tree, we do not propagate any points-to information down to the children in the points-to tree. So to apply the CollapseDepth correctly on the TaintTree built during deep read using a PointsToTree, we need to iterate on the points-to tree from leaf to root. This diff adds a `visit_postorder()` helper to the abstract tree domain to do so.

Reviewed By: arthaud

Differential Revision: D70432277

fbshipit-source-id: 50ce4fe35b1401b68dfb110f9cfadb636beeaa5f