Add nested class tests for alias on propagations
Summary:
In java/kotlin, nested class (inner classes, local classes, anonymous classes) can access to the enclosing scope. This access is provided through synthetic members added to the nested classes. The  reference to the enclosing scope (parent class' "this" pointer or for local and anonymous classes, any captured local variables or method parameters) are implicit arguments to the constructor of the nested class.

This diff adds tests to demonstrate the cases of aliasing the enclosing scope through `this$n` access paths.
In the following diffs, we will correctly resolve these aliases to the enclosing scope and find the FNs.

Reviewed By: yuhshin-oss

Differential Revision: D71290537

fbshipit-source-id: 3c960b7b7eab873d80e99e5be536c24eadff835a