Update the deep_read to account for collapse depths from propagation models
Summary:
This diff stack is adding support for aliasing propagations. This means that the points-to-tree can now contain collapse-depths other than `0` (i.e. always collapse, added on widening) and `oo` (i.e. no-collapse, for intraprocedural analysis (iputs/igets)).

This diff updates the `deep_read()` implementation to handle this accurately instead of applying the collapse-depth directly on the final result taint.

Reviewed By: arthaud

Differential Revision: D71346974

fbshipit-source-id: e09f711f4c55e42c9aceb19c3408813213a51f17