Handle specific aliasing propagations for .this$n access paths
Summary:
Nested classes (local, inner, or anonymous) provide access to the members/local variables/method arguments of the enclosing scope by storing a reference to them in "synthetic fields". These fields are implicitly passed to the constructors of the nested classes and any access to the enclosing scope is implicitly accessed through them.
They are called synthetic because they cannot be referenced in the source code and hence follow naming convention including invalid characters like `$` (similar to the internal names of anonymous classes). These fields alias the memory location of the parent scope and are a major source of false negatives in mariana trench.

This diff implements a v1 for apply aliasing to propagations. Specifically, this diff only handles the aliasing propagations of the form: `Argument(m) -> Argument(0).this$n`. This is the most common form of aliasing propagation as it is applicable to methods of all nested classes (inner, local, or anonymous).
- `maybe_get_aliasing_propagation_source_memory_location()` processes the input paths and the corresponding memory locations to further limit applying aliasing to:
  - input AccessPath: is argument and does not have any path elements
  - input path points to a singleton root memory location
- `maybe_get_aliasing_output_path_element()` processes the output path to limit applying aliasing to:
  - output root is `Argument(0)`
  - output path is a single path element matching the naming convention `this$n`

For propagations that are considered valid for aliasing as above: we add an alias i.e. update the points-to-tree at the output path to point-to the source memory location of the input.

Following diffs will build on this to handle all aliasing propagations through synthetic fields.

Reviewed By: yuhshin-oss

Differential Revision: D71290536

fbshipit-source-id: 447ceb9866120168e1c241d60fc7ffc2d5e54466