Move Registry::from_sharded_models_json into CachedModelsContext
Summary:
We have `read_[class_intervals|class_hierarchies|overrides]()` in `CachedModelsContext` but for some reason, parsing the models was done in `Registry`. This moves the model parsing into `CachedModelsContext`.

NOTE: `CachedModelsContext` will be removed later and the `read_*` goes back into their respective classes. In the current intermediate state, though, we'll have it live in `CachedModelsContext` for consistency.

Reviewed By: anwesht

Differential Revision: D71847095

fbshipit-source-id: f3d069f1b5c92c829dc8836c4999bee6d3584bc4