Registry: Use const Registry* for cached registry (instead of std::optional)
Summary:
From an earlier diff review comment: `std::optional` could result in inadvertent copies, so use `const Registry* MT_NULLABLE` instead.

Note that `Registry` is currently move-only, so it's not possible to copy it, but I do agree `const Registry*` is nicer anyway.

We do need to be-careful in the model generators that the underlying pointer is valid throughout the duration of the model generator though

Reviewed By: anwesht

Differential Revision: D72197304

fbshipit-source-id: e626060f40ec9c06fa23ac8c6b9d73f8f13dffbd