Fix JsonValidation potential dangling references
Summary:
Building with GCC shows a lot of warnings on JsonValidation methods as follows:
```
/home/runner/work/mariana-trench/mariana-trench/source/FieldModel.cpp:143:66: warning: possibly dangling reference to a temporary [-Wdangling-reference]
  143 |        JsonValidation::null_or_array(value, /* field */ "sources")) {
```

On close inspection, this might happen when we directly access a field as `json[field]` and return a reference to it iff `field` does not exist in the original `json` object. In this case, we return a reference to a temporary `Json::Value` object. Let's fix it. This diff adds:
- the methods that return a reference to a null/empty object returns a reference to `Json::Value::nullSingleton();` instead of the temporary.
- also adds `validate()` method with `field` as argument which does a `isMember()` check for it. Use this when we require the field to be present

Differential Revision: D72485842