Update RulesCoverage to match FilesCoverage style
Summary:
Similar to `FilesCoverage`, this updates the existing `RulesCoverage` to be stand-alone and called directly from MarianaTrench.cpp (instead of via `registry.dump_rule_coverage_info()`).

No functional changes. Mainly API and unit test changes. Basically `RulesCoverage::create()` takes in the `Registry` and `Rules`, both of which will expose whatever is necessary for computing rules coverage. Done largely for the sake of being consistent with `FilesCoverage`.

This does make it more annoying to unit test since the registry now needs to be constructed correctly (as opposed to just passing in a list of sources/sinks/transforms).

Reviewed By: anwesht

Differential Revision: D72402694

fbshipit-source-id: 6fda720d9ef754344508f357144d7252e6564213