FilesCoverage: Consider files without any implementation as covered
Summary:
As the title says. Files without any implementation is reported as covered by `FilesCoverage`. We use Redex's `is_abstract()` to identify "no implementation". All methods in the file must pass the `is_abstract()` check for the file to be considered covered.

NOTE: Since `is_abstract()` applies to methods only, files with no methods (e.g. enums, static fields only, etc.) are not taken into account. Also, we are using `Positions` to obtain the fully resolved file path, and it only tracks method-to-filename mappings. More needs to be done if we want to support the other cases. Since there aren't many of such files compared to interface-only files, this is ignored for now.

Reviewed By: anwesht

Differential Revision: D72092170

fbshipit-source-id: 7b49772593324fdef428cc2acdfcffa1c58521a7