Add FilesCoverage class
Summary:
Will be adding the computation of interface/enum files to `FilesCoverage` in the next diff.

Today, file coverage info is computed in `Registry` because it is based solely on the models. This will no longer be the case once we include interface/enum-only files. Those are computed solely based on the loaded methods/scope.

Reviewed By: anwesht

Differential Revision: D71989742

fbshipit-source-id: eb730235d8db3d10c39b061edeb0a2910f4d5e87