Fix apply generations for artificial calls
Summary:
Problem:
We infer generations on arguments through artificial calls. However, we do this in a loop while checking for flows through artificial calls. This is problematic and leads to non-determinism and FPs depending on the iteration order.

Solution:
This diff extracts the `apply_artificial_calls_generations()` and applies it only after we are done checking flows due to artificial calls (same as for real calls).

Reviewed By: arthaud

Differential Revision: D72736121

fbshipit-source-id: d8582106cc57b4a9747d85ef6e0c9e5fd2b5cb76