Add rule that detects TrustedCaller bypasses
Summary: We'd like to be alerted when Android Vanilla APIs are used to validate which app is calling our app. While not always a vulnerability we generally prefer our SAF APIs to be called such as TrustedCaller and such verifications may lead to potential bypasses such as accessibleByAnyApp callsites still passing the validation.

Differential Revision: D73030181

fbshipit-source-id: e2a66ce8a9017ddb3f243d6f60a82eab3a8ec555