Do no generate via_type|value features for invalid ports in models
Summary:
Currently we add `via_[value|type]_of` to a method's model even if the ports are invalid. We log this during model generation but since we still add this to the model, we keep trying to materialize the features at all callsites of this method which is unnecessary and also pollutes the logs.

This diff filters out the `via_[value|type]_of` of features when adding a model for the method based on the number of arguments of the method since we allow broad model generators where `via_[value|type]` features may be applicable to some but not all of the matched methods).  The invalid ports will only be logged once during model generation.

Reviewed By: arthaud

Differential Revision: D74920256

fbshipit-source-id: 4bd391fafdf86b934fea357f79cb6ec23c668960