Add an integration test for graph-like lifecycle transitions
Summary:
Add an integration test to show a possible FN cases with our current linear
lifecycle state transitions and as an preparation for the support on graph-like
lifecycle transitions that will come in later diffs.

For the possible FN, consider the following case:

```
public class MainActivity extends Activity {
  private Object mSource;

  Override
  protected void onCreate(Bundle savedInstanceState) {
    // Expect issue
    Origin.sink(mSource);
  }

  Override
  protected void onStart() {
    mSource = Origin.source();
  }
}
```

Our existing linear lifecycle would only consider the lifecycle state
transition `onCreate -> onStart`, which will not report an issue. However, in
real life, the transition is not necessarily linear [1] --- there could exist a
transition `onStart -> onCreate`, and this will become a false-negative.

[1]: https://developer.android.com/guide/components/activities/activity-lifecycle

Reviewed By: anwesht

Differential Revision: D78575230

fbshipit-source-id: a5b32a57b382d8fc414ac51e623066c0b5870095