Merge lifecycle method generation log messages with method code dumps
Summary:
`LifecycleMethod::create_dex_method` is invoked from a `sparta::work_queue`,
which by default runs multiple threads in parallel. This parallelism causes
potential interleaving of log messages. In particular, it may split the
"Created life-cycle method `{}` ..." message from the subsequent "Generated
method body:" message even though the logger is called back-to-back, making it
hard to figure out which method the method body belongs to. To fix this, let's
merge the two logs into a single call to the logger.

Reviewed By: anwesht

Differential Revision: D78700209

fbshipit-source-id: cfefe774dfa703da45d8ad9a1f45db69ead4d072