Sort arguments of lifecycle wrappers to ensure consistent ordering between linear and graph based lifecycles
Summary:
We are seeing changes in generated models following D78380996, where the order
of arguments in the generated lifecycle wrapper is different for graph-based
lifecycles. This is because the definition of graph-based lifecycles in the
JSON DSL is now a dict (node-name -> node-definition) rather than a strictly
ordered array. The difference in the generated models could potentially change
our issue handles in the SAPP UI, leading to difficulties in triaging.
Therefore, sort the lifecycle wrapper arguments based on their type name to
ensure a consistent ordering between linear- and graph-based lifecycles.

Reviewed By: anwesht

Differential Revision: D79311172

fbshipit-source-id: 4f9cb5c1ab9f885fefc6fb9c1e4cabbc7713f41f