Allow specifying return types of shimmed methods as receivers in shim config
Summary:
As the title goes, this diff implements the parsing needed for handling return
types of shimmed methods as receivers. The actual shim instantiation logic will
come in followup diffs.

P.S. Refactored the `ReceiverInfo` class to directly store a `Root` rather than
a parameter index since that makes it easier to deal with receivers on returns.

Reviewed By: anwesht

Differential Revision: D80306712

fbshipit-source-id: 6b7999fc2fb193daebc9e771a843ab6ba7b55d36