new sources IntentHandlers
Summary: This diff introduces new sources for IntentHandlers in Mariana Trench. Specifically, it adds a new model generator for IntentHandlers that generates sources for methods that extend the IntentHandler interface and have a signature matching the getIntent method. The new sources are generated using the "ActivityUserInput" kind and are marked with the "via-public-dfa-scheme" feature. This diff also includes a unit test for the new model generator.

Reviewed By: anwesht

Differential Revision: D80887137

fbshipit-source-id: 69bc695951b6a00a59583f96101da829f4fb594d