Make ShimParameterMapping::MapType a map from Root to Root
Summary:
Previously this maps from a Root representing a parameter of the shim target to
a parameter index of the shimmed method. But with the introduction of shim
receivers on return ports, argument 0 (i.e., `this`) of the shim target can now
map to the return port of the shimmed method.

Therefore, refactor the map to make it a Root-to-Root mapping so it is easier
to support shim receivers on return ports.

Reviewed By: anwesht

Differential Revision: D80828768

fbshipit-source-id: d12b1dbbeb44e001cb7402fd8994c607980e835c