Generate invocations to shim targets with receivers on return ports of shimmed methods
Summary:
As title, this diff sets up the correct registers (i.e. `RESULT_REGISTER`) for
return values for our call graph construction code to find the type and value
of the return, and generate the invocation to shim targets. It also extends the
existing shim E2E test with cases involving receivers on return ports.

Reviewed By: anwesht

Differential Revision: D80718322

fbshipit-source-id: 9484b9a415417d4cf77716fa2eefe7f210d0472a