Refine types of return values of invoke instructions using global type analysis
Summary:
Followup of the type analysis on the return register of method invocations. The
local type analysis gives us nothing more than the return type declared in the
method prototype and its local extends. This diff uses the type information
from the existing global type analysis to further refine the type of the return
register.

Reviewed By: anwesht

Differential Revision: D81286330

fbshipit-source-id: 6c6628592eac43b3b4af53e898423fa743d55cf7