Support lifecycle shim targets on return instance
Summary:
Following the basic support of shims on return instances, this diff extends the
support to include the lifecycle wrappers as shim targets. In other words, we
can now generate calls to lifecycle wrappers with return value of the
shimmed-method acting as the instance object and capture more taint flows.

Reviewed By: anwesht

Differential Revision: D82066975

fbshipit-source-id: e42c5d50850d6bfeade2a6b1ff06a9dfa3112a2b