Update Multi Source documentation
Summary:
There aren't "as many sinks as there are sources", there are as many **partial_labels** as there are sources.

Change from this:
 {F1981940389}

To this:

 {F1981940363}

In the rule below, it went on to show 2 sources and 1 sink, so I think the wording there was misleading.
{F1981940414}

I cleaned it up and broke it into bullets, linking the port documentation. I think it's much easier to understand the requirements for each partial sink now.

Reviewed By: anwesht

Differential Revision: D82319848

fbshipit-source-id: 84b16b28bfb0dbed630c9a581ce246112d50f1c9