Add test case for FP where base class implementations should not be called in lifecycle wrappers
Summary:
This diff adds a test case to capture possible FPs triggered when the base
class implementation of a lifecycle method is not supposed to be called. For
example, a method `onXXX()` may only be invoked if `canXXX()` returns true in
the subclass, which indicates that the subclass has overriden `onXXX()`. In
this case, the invocation of `onXXX` from the base class indicates
implementation errors and should not happen in real world apps.

Reviewed By: anwesht

Differential Revision: D82354345

fbshipit-source-id: c59888f9ff70a220e66ce81424c4148afbd18779