Allow skipping base class implementation of lifecycle methods
Summary:
Introduce a new key `"skip_base_implementation"` in the lifecycle config that
can be applied to individual lifecycle methods. If a lifecycle method is marked
with `"skip_base_implementation"` and is resolved to the base class
implementation, it will not be called in the generated lifecycle wrapper. The
diff also includes small cleanups of the existing code.

Currently only implemented for linear lifecycles, support for lifecycle graphs
will come in the future.

Reviewed By: anwesht

Differential Revision: D82285302

fbshipit-source-id: 5085de8ff384cfc8f20d3a2552ab02e9a2562d8d