Support "skip_base_implementation" in lifecycle graphs
Summary:
As the title suggests, this follow-up diff completes the handling of
`"skip_base_implementation"` by adding support for lifecycle graphs. In
lifecycle graphs, if we need to skip the base class implementation of a method,
we skip generating the invoke instruction in the basic block associated with
the lifecycle graph node to which the method belongs in the configuration. Note
that doing so may result in basic blocks without invoke instructions, but the
generated lifecycle wrapper is still valid.

Reviewed By: anwesht

Differential Revision: D82596054

fbshipit-source-id: c4e2f04077cc412cca3657c68b20471c962ca8c2