Open sourcing AndroidxNavigationUserInput
Summary:
`AndroidxNavigationUserInput` is generic to all Android, so this makes sense to add this to the open-source configuration for anyone to use.

https://developer.android.com/jetpack/androidx/releases/navigation

Arguments can be added to the navigation via `NavController`'s `.navigate()`:
```
navController.navigate("user1234")
```

See [deeplinks with Androidx Navigation](https://developer.android.com/develop/ui/compose/navigation#deeplinks).

- These arguments can be fetched later from the `NavController`.
- `BackStackEntry`s are stored in a stack in the NavController.
- Fetching a BackStackEntry from the stack and then calling`getArugments` (invoked by simply `.arguments` with Kotlin) will fetch the Bundle. This is what Mariana Trench is tracking as the source.
```
navBackStackEntry.arguments?.getString("id")
```
In order for this to be "real" attacker input, it most likely needs to flow through a `.getString` or similar bundle function, which can be configured with transformers or features.

Reviewed By: GerbenJavado

Differential Revision: D83403771

fbshipit-source-id: ddd91da2bc9ad7c78a861957f913bec592c73297