Stop using exception handling for control flow
Summary:
The code where we query field models in `Registry::get()` uses `try-catch` for
control flow, i.e., the code relies on `ConcurrentMap::at()` throwing an
exception to handle the case where the field model is not found and return a
default model. This is not ideal given that exception handling in C++ is
expensive [1]. Therefore, refactor the code to remove the expensive exception
handling logic and implement it using `ConcurrentMap::get()` instead.

[1]: https://itanium-cxx-abi.github.io/cxx-abi/abi-eh.html

Reviewed By: anwesht

Differential Revision: D85734478

fbshipit-source-id: 823832a62a9fdcd5e7dc189542006a05b33531ee