Sanitize source as transform sinks when callkind is IntentRouting
Summary:
In mariana-trench we use call-chain-exploitability to determine if the downstream issue's root callable is accessible from an exported Activity.

Currently, we have a false positive where a call-chain-exploitability trace is found event through Activity launchers. But when an Activity is launched via an intent routing shim, the exploitability call-chain should be reset (i.e. it should sanitize the propagation of `SourceAsTransform[<sink>]`) as the launched Activity may or may not be exported.

This diff when the `callee` is found to have a call-effect-sink on the `CallEffectExploitability` port, we additionally check the `CallKind` of the `callee` and only  propagate the `SourceAsTransform[<sink>]` to the backward analysis if the call kind is *not* `CallKind::IntentRouting`.

 ---

Reviewed By: jinghao-jia

Differential Revision: D87474060

fbshipit-source-id: 03dea08a635018c80d428e2cf34c69e18dc9df22