Fix and productionize InputStream
Summary:
This sink catches many methods such as:
- Ljava/io/FileInputStream;.<init>:(Ljava/io/File;)V
- Ljava/io/ByteArrayInputStream;.<init>:([B)V

Which do not suffer from path traversal risk where this does apply to
- Ljava/io/FileInputStream;.<init>:(Ljava/lang/String;)V

We may want to find later on where direct bytes are getting read into an object but that would need a distinct rule investigation. For now extending this on top of the File categories that also concern themselves with reading/writing to Files in the appsandbox.

Reviewed By: arthaud

Differential Revision: D88477389

fbshipit-source-id: b50c6f58d620b69778502711c9d23e9975b45db0